fx-project/backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
pymongo
pandas
numpy
scikit-learn
requests